# web/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "AppleName": {
        "properties": {
          "firstName": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstname"
          },
          "lastName": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastname"
          }
        },
        "title": "AppleName",
        "type": "object"
      },
      "AppleUser": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AppleName"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AppleUser",
        "type": "object"
      },
      "AuthTokenResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "is_new_user": {
            "title": "Is New User",
            "type": "boolean"
          },
          "refresh_expires_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Expires In"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "required_consents": {
            "items": {
              "$ref": "#/components/schemas/RequiredConsent"
            },
            "title": "Required Consents",
            "type": "array"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "is_new_user",
          "user"
        ],
        "title": "AuthTokenResponse",
        "type": "object"
      },
      "AuthorizeResponse": {
        "additionalProperties": false,
        "properties": {
          "authorize_url": {
            "title": "Authorize Url",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "authorize_url",
          "state",
          "expires_in"
        ],
        "title": "AuthorizeResponse",
        "type": "object"
      },
      "Body_apple_notifications_api_v1_auth_apple_notifications_post": {
        "properties": {
          "payload": {
            "maxLength": 8192,
            "title": "Payload",
            "type": "string"
          }
        },
        "required": [
          "payload"
        ],
        "title": "Body_apple_notifications_api_v1_auth_apple_notifications_post",
        "type": "object"
      },
      "ConsentItem": {
        "additionalProperties": false,
        "properties": {
          "agreed": {
            "title": "Agreed",
            "type": "boolean"
          },
          "type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "version": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "type",
          "version",
          "agreed"
        ],
        "title": "ConsentItem",
        "type": "object"
      },
      "ConsentListResponse": {
        "additionalProperties": false,
        "properties": {
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentState"
            },
            "title": "Consents",
            "type": "array"
          }
        },
        "required": [
          "consents"
        ],
        "title": "ConsentListResponse",
        "type": "object"
      },
      "ConsentState": {
        "additionalProperties": false,
        "properties": {
          "agreed": {
            "title": "Agreed",
            "type": "boolean"
          },
          "type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "type",
          "version",
          "agreed"
        ],
        "title": "ConsentState",
        "type": "object"
      },
      "ConsentSubmitRequest": {
        "additionalProperties": false,
        "properties": {
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentItem"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Consents",
            "type": "array"
          }
        },
        "required": [
          "consents"
        ],
        "title": "ConsentSubmitRequest",
        "type": "object"
      },
      "ConsentType": {
        "enum": [
          "terms_of_service",
          "privacy_policy",
          "age_over_14",
          "marketing",
          "quality_log_collection"
        ],
        "title": "ConsentType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogoutRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "title": "LogoutRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "maxLength": 256,
            "minLength": 32,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "RequiredConsent": {
        "additionalProperties": false,
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "type",
          "version",
          "url"
        ],
        "title": "RequiredConsent",
        "type": "object"
      },
      "SocialLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "apple_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AppleUser"
              },
              {
                "type": "null"
              }
            ]
          },
          "code": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "code_verifier": {
            "maxLength": 128,
            "minLength": 43,
            "pattern": "^[A-Za-z0-9\\-._~]+$",
            "title": "Code Verifier",
            "type": "string"
          },
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentItem"
            },
            "maxItems": 10,
            "title": "Consents",
            "type": "array"
          },
          "device_info": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Info"
          },
          "handoff": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handoff"
          },
          "redirect_uri": {
            "maxLength": 2048,
            "title": "Redirect Uri",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "required": [
          "code_verifier",
          "redirect_uri"
        ],
        "title": "SocialLoginRequest",
        "type": "object"
      },
      "SocialProvider": {
        "enum": [
          "kakao",
          "google",
          "naver",
          "apple"
        ],
        "title": "SocialProvider",
        "type": "string"
      },
...
      "TokenRefreshResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_expires_in": {
            "title": "Refresh Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "refresh_expires_in"
        ],
        "title": "TokenRefreshResponse",
        "type": "object"
      },
...
      "UserProfile": {
        "additionalProperties": false,
        "properties": {
          "default_target_lang": {
            "title": "Default Target Lang",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_private_email": {
            "title": "Is Private Email",
            "type": "boolean"
          },
          "linked_providers": {
            "items": {
              "$ref": "#/components/schemas/SocialProvider"
            },
            "title": "Linked Providers",
            "type": "array"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "ui_locale": {
            "title": "Ui Locale",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "email",
          "is_private_email",
          "status",
          "ui_locale",
          "default_target_lang",
          "plan",
          "linked_providers"
        ],
        "title": "UserProfile",
        "type": "object"
      },
      "UserResponse": {
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "required": [
          "user"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserStatus": {
        "enum": [
          "active",
          "suspended",
          "pending_consent"
        ],
        "title": "UserStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TalkSoft API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/apple/notifications": {
      "post": {
        "operationId": "apple_notifications_api_v1_auth_apple_notifications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_apple_notifications_api_v1_auth_apple_notifications_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Apple Notifications Api V1 Auth Apple Notifications Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apple Notifications",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/authorize/{provider}": {
      "get": {
        "operationId": "authorize_api_v1_auth_authorize__provider__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code_challenge",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{43}$",
              "title": "Code Challenge",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code_challenge_method",
            "required": true,
            "schema": {
              "const": "S256",
              "title": "Code Challenge Method",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "schema": {
              "maxLength": 2048,
              "title": "Redirect Uri",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authorize",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/consents": {
      "get": {
        "operationId": "list_consents_api_v1_auth_consents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Consents",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "submit_consents_api_v1_auth_consents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Consents",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/link/{provider}": {
      "post": {
        "operationId": "link_provider_api_v1_auth_link__provider__post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link Provider",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login/{provider}": {
      "post": {
        "operationId": "login_api_v1_auth_login__provider__post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LogoutRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}